A website where people can post stories and reviews on things they like and care about.
StoryTime basically is a website that people can use to post stories and
reviews on things they like and care about. I built a sample version of
StoryTime for a project at the university.

People can register for an account and post stories or reviews.
Anyone can comment, like, and share stories or reviews posted.